* commit '6795e9c783242b31188d76c5186dfe0751f1731a':
  AArch64: Use long for pointers in SurfaceSession class

+++ core/java/android/view/SurfaceSession.java
@@ -24,11 +24,11 @@ package android.view;
  */
 public final class SurfaceSession {
     // Note: This field is accessed by native code.
-    private int mNativeClient; // SurfaceComposerClient*
+    private long mNativeClient; // SurfaceComposerClient*
 
-    private static native int nativeCreate();
-    private static native void nativeDestroy(int ptr);
-    private static native void nativeKill(int ptr);
+    private static native long nativeCreate();
+    private static native void nativeDestroy(long ptr);
+    private static native void nativeKill(long ptr);
 
     /** Create a new connection with the surface flinger. */
     public SurfaceSession() {
